Bump odlparent->6.0.0,mdsal->5.0.3

+++ aclservice/impl/src/test/java/org/opendaylight/netvirt/aclservice/tests/AclServiceTestBase.java
@@ -65,6 +65,7 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.elan.rev150602.elan
 import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.elan.rev150602.elan.interfaces.ElanInterface;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.elan.rev150602.elan.interfaces.ElanInterfaceBuilder;
 import org.opendaylight.yangtools.yang.binding.InstanceIdentifier;
+import org.opendaylight.yangtools.yang.common.Uint64;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -129,7 +130,7 @@ public abstract class AclServiceTestBase {
     }
 
     private InterfaceInfo newInterfaceInfo(String testInterfaceName) {
-        InterfaceInfo interfaceInfo = new InterfaceInfo(BigInteger.valueOf(789), "port1");
+        InterfaceInfo interfaceInfo = new InterfaceInfo(Uint64.valueOf(BigInteger.valueOf(789)), "port1");
         interfaceInfo.setInterfaceName(testInterfaceName);
         return interfaceInfo;
     }
